Move main update code into separate function and loop forever
We now loop forever calling update_display with a 10 minute delay. This
more easily lets us set this up as a service on the Raspberry PI, so it
can automatically start on boot.

We'll still need to get rid of all the unwrap() calls so we don't blow up
if any GitHub API calls fail, and the 10 minute delay shouldn't be
hard-coded, but these are things that can be handled later.