(#10) Add support to handle an input of zero
This commits adds support to the main function that maps the leds to
handle a total of zero. If this is the case, we don't want to fail, we
simply want to report a vector with elements that add up to zero, rather
than try to map it so that it adds up to 16.

This commit also breaks up the tests so that each test vector is tested
individually.